Re-add live stream handling
This change adds a workaround to detect if a stream is live or not.
While twitch seemed to remove any indication of ongoing streams they
also wanted to spare bandwidth by not adding a link to the thumbnails
while the stream is on. So there it is, we now detect live streams by
checking the thumbnail if it has a link, if not => it's live.

Sadly this can be broken by twitch any time. I give it 6 months.